feat: refatorar lógica de mensagens citadas, movendo a construção para uma função separada
